revision#2 after 2nd feedback

+++ src/processor.cpp
@@ -7,10 +7,10 @@ using std::string;
 using std::vector;
 
 // TODO: Return the aggregate CPU utilization
-float Processor::Utilization() {
+double Processor::Utilization() {
   long Active = LinuxParser::ActiveJiffies();
   long Idle = LinuxParser::IdleJiffies();
-  float cpuUtilization = Active / double(Active + Idle);
+  double cpuUtilization = Active / double(Active + Idle);
 
   return cpuUtilization;
 }
